Change Elo values from decimal to integer, and set default value
Adjusted the Elo value types from decimals to integers in the match controller and player model files. Default Elo value set to 1500 when a new user is created. This creates a more intuitive value type for Elo ratings and establishes a standard baseline for new users.

+++ Server/Models/Player.cs
@@ -4,7 +4,7 @@ public class Player
 {
     public Guid Id { get; set; }
     public string NickName { get; set; }
-    public decimal Elo { get; set; }
+    public int Elo { get; set; }
     public Uri? ImageUrl { get; set; }
     public string UserId { get; set; }
     public ApplicationUser User { get; set; }
